// src/app/Model/Project.ts
export interface Project {
  id?: number;
  clientProjectId: string;
  projectName: string;
  startDate: string;
  expectedEndDate: string;
  leadByEmpId: string;
  completedDate: string;
  contactPerson: string;
  contactPersonContactNo: string;
  totalEmpWorking: string;
  projectCost: string;
  projectDetails: string;
  contactPersonEmailId: string;
}